probability DFS : 2 n 4 fail

+++ Hackerrank/dfsprobreach.cpp
@@ -148,7 +148,7 @@ int main()
 
 		}
 
-		cout << allprobs << endl;
+		cout << (int)allprobs << endl;
 
 	}
 	return 0;
